requirements.txt
opencv-python
torch
torchvision
Pillow
pytorch-lightning
timm
nltk
numpy
